Move receipt endpoint to http handler
